moved the pseudo mesh to seperate script and precomputed it
pseudo mesh for hands now gets precomputed in seperat script and safed in the inspector, should be moved to a file in the future. this way we dont need to copmute the pseudo mesh at start. also made the other factor start when hands are close to hmd so people that have the hands at their hips when it starts dont end up with the particles on the floor and this way i removed the default 2 sec delay for people to put their hands up