Replaced default inspector with one that can safe and load presets
added a new script that can safe and load the variable values of theotherfactor that define a "preset". they can be accessed and manged through a custom inspector but it is also possible to siwtch between created presets using pinch. This allows for a comparison between different settings without the need to remove the headset. todo: display all values on the canvas that floats in front of the camera