added realtime mirror option
added the option to do a real time mirroring of the hand joints directly inside the other fator. also made the mirror position and rotation determined by the direction and position of the camera when hands trigger the emit of particles so it is always in front of the user at start. made some minor adjustments to the attraction calculation that keeps things in the proper ranges with mroe control