Improvements
main change is in the way the speed of the hands is affecting the attraciton. so far i used the speed of the individual joints, but since particles wher not always attracted to the same joints on both hands, the behaviour didnt look good. now there is one joint on each hand which determines the speed for attraction for the whole hand. the specific joint can be important, since we want to capture the rotation of the hand as well. also included values to clamp the speed according to a range of distance traveled between updates. further i adjusted the way in which the mesh indices get traversed to create a nicer spread of particles, which leads ot better behaviour at slow speeds as well as more defined hands. slight adjustment to the way the mesh path is traversed. also added a rotation option to the torus